requirements.txt
# --- Core SERP/proxy integration ---
serpapi>=0.1.5,<0.2
scrapfly-sdk>=0.8.5,<0.9
scraperapi-sdk>=1.1.0,<2.0

# --- HTTP & parsing ---
requests>=2.31.0,<3.0
beautifulsoup4>=4.12.0,<5.0
lxml>=4.9.3,<5.0
extruct>=0.17.0,<0.19
soupsieve>=2.4.1,<3.0

# --- AI & templating ---
openai>=1.3.0,<2.0
jinja2>=3.1.2,<4.0
python-dotenv>=1.0.0,<2.0

# --- Utilities & scheduling ---
tqdm>=4.66.0,<5.0
apscheduler>=3.10.1,<4.0
selenium>=4.15.2,<5.0
webdriver-manager>=4.0.0,<5.0
pandas>=2.1.0,<3.0

# --- Scraping robustness ---
tenacity>=8.2.3,<9.0
requests-toolbelt>=0.10.1,<1.0

# --- Testing & CLI experience ---
pytest>=7.4.0,<8.0
requests-mock>=1.11.0,<2.0
rich>=13.7.0,<14.0
